feat: refactor AuthService and tokenInterceptor to core directory and update import paths
