+++ src/app/app.config.ts
@@ -1,7 +1,7 @@
 import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
 import { provideRouter } from '@angular/router';
 import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptors } from '@angular/common/http';
-import { tokenInterceptor } from './services/token.interceptor'; // ← correcto
+import { tokenInterceptor } from './core/services/token.interceptor'; // ← correcto
 
 import { routes } from './app.routes';
