+++ src/app/components/vet-clinic-login/login.component.ts
@@ -8,7 +8,7 @@ import {
   Validators
 } from '@angular/forms';
 import { Router, RouterLink } from '@angular/router';
-import { AuthService } from '../../services/auth.service'; // Asegúrate de que la ruta sea correcta
+import { AuthService } from '../../core/services/auth.service'; // Asegúrate de que la ruta sea correcta
 
 @Component({
   selector: 'app-login',
