+++ src/main.ts
@@ -4,7 +4,7 @@ import { AppComponent } from './app/app.component';
 import { registerLocaleData } from '@angular/common';
 import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptors } from '@angular/common/http';
 import localeEs from '@angular/common/locales/es';
-import { tokenInterceptor } from './app/services/token.interceptor';
+import { tokenInterceptor } from './app/core/services/token.interceptor';
 bootstrapApplication(AppComponent, appConfig)
   .catch((err) => console.error(err));
